Improve transformation of inputs passed to Function.create()
Currently we only convert `URLPath` instances into URL strings if passed
directly as an input parameter. We want to ensure that we convert all
instances of `URLPath`s provided otherwise we get JSON encoding errors.

The most common use of this is passing a list of files that were output
from another model.

This PR attempts to transform the most common Python data structures
into either lists or dicts and in doing so extracts the underlying URL
value from any `URL` path instances.

This might be better implemented as a custom JSON encoder.